stock-crawler/requirements.txt
python-dotenv==1.0.1
requests==2.31.0
psycopg2-binary==2.9.9
schedule==1.2.1
yfinance==0.2.50
